# openapi.work.yaml
openapi: 3.0.1
info:
  title: Deployer API
  version: 0.3.0
  contact:
    name: Juha Itäleino
    email: jsilvanus@gmail.com
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: DEPLOYER_ADMIN_TOKEN
    AppKey:
      type: apiKey
      in: header
      name: X-APP-KEY
  schemas:
    VersionResponse:
      type: object
      properties:
        id:
          type: string
        version:
          type: string
        source:
          type: string
          description: git|image|npm|pypi
        ref:
          type: string
          description: commit SHA or image digest
        deployedAt:
          type: integer
          description: epoch seconds
        deployedBy:
          type: string
        notes:
          type: string

    VersionDiffResponse:
      type: object
      properties:
        deployed:
          $ref: '#/components/schemas/VersionResponse'
        latest:
          $ref: '#/components/schemas/VersionResponse'
        upgradeAvailable:
          type: boolean

    PaginatedVersionsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VersionResponse'
        nextCursor:
          type: string

    RegistryTestRequest:
      type: object
      properties:
        provider:
          type: string
          description: docker|npm|pypi|git
        target:
          type: string
          description: image name, package name, or git url
        credentials:
          type: object
          additionalProperties: true

    RegistryTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        tags:
          type: array
          items:
            type: string

    ShutdownRequest:
      type: object
      properties:
        action:
          type: string
          enum: [stop, restart, graceful, destroy]

    CachePurgeRequest:
      type: object
      properties:
        target:
          type: string
        paths:
          type: array
          items:
            type: string

    OperationResponse:
      type: object
      properties:
        operationId:
          type: string
        status:
          type: string

security:
  - AdminToken: []

paths:
  /apps/{appId}/version:
    get:
      summary: Get current deployed version for an app
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '200':
          description: Current version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'

  /apps/{appId}/version/latest:
    get:
      summary: Compare deployed vs upstream latest (cached by default)
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: refresh
          in: query
          schema:
            type: boolean
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '200':
          description: Version comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDiffResponse'

  /apps/{appId}/registry/test:
    post:
      summary: Test registry credentials or access for an app (transient)
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistryTestRequest'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryTestResponse'

  /apps/{appId}/versions:
    get:
      summary: List historical versions/deployments for an app (paginated)
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '200':
          description: Paginated versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVersionsResponse'

  /apps/{appId}/shutdown:
    post:
      summary: Stop/restart/shutdown an app
      description: |
        Non-destructive actions (`stop`, `restart`, `graceful`) are allowed with an app-scoped API key.
        Destructive action `destroy` requires an admin token and is gated by server config.
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShutdownRequest'
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '202':
          description: Operation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'

  /apps/{appId}/schedules:
    get:
      summary: List schedules for an app
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

    post:
      summary: Create a schedule for an app
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created

    delete:
      summary: Delete a schedule by id (use body or query to specify)
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      responses:
        '204':
          description: Deleted

  /apps/{appId}/cache/purge:
    post:
      summary: Purge app-related caches
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: string
      security:
        - AdminToken: []
        - AppKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CachePurgeRequest'
      responses:
        '202':
          description: Purge scheduled/executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'